requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
dbus-fast==3.1.2
idna==3.11
microdot==2.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
tinydb==4.8.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
